# backend/openapi.yaml
openapi: 3.0.3
info:
  description: API para Sistema de Gestión de Condominios
  title: Casitas Teto API
  version: 1.0.0
servers:
- url: /
...
paths:
...
  /usuarios:
    get:
      parameters:
      - description: ID del condominio
        in: query
        name: condominioId
        required: false
        schema:
          type: integer
      - in: query
        name: rol
        schema:
          $ref: "#/components/schemas/RolUsuario"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_usuarios_get_200_response"
          description: Lista de usuarios
      summary: Listar usuarios
      tags:
      - Usuarios
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
          description: Usuario creado
      summary: Crear usuario
      tags:
      - Usuarios
  /usuarios/{usuarioId}:
    get:
      parameters:
      - description: ID del usuario
        in: path
        name: usuarioId
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
          description: Detalles del usuario
      summary: Obtener usuario
      tags:
      - Usuarios
    put:
      parameters:
      - description: ID del usuario
        in: path
        name: usuarioId
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioInput"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
          description: Usuario actualizado
      summary: Actualizar usuario
      tags:
      - Usuarios
  /usuarios/{usuarioId}/deshabilitar:
    patch:
      parameters:
      - description: ID del usuario
        in: path
        name: usuarioId
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: Usuario deshabilitado
      summary: Deshabilitar usuario
      tags:
      - Usuarios
...
components:
...
  schemas:
...
    Usuario:
      example:
        apellido: apellido
        fechaCreacion: 2000-01-23T04:56:07.000+00:00
        condominioId: 6
        id: 0
        ultimoAcceso: 2000-01-23T04:56:07.000+00:00
        nombre: nombre
        email: email
        rol: SUPER_ADMINISTRADOR
        activo: true
      properties:
        id:
          title: id
          type: integer
        email:
          format: email
          title: email
          type: string
        nombre:
          title: nombre
          type: string
        apellido:
          title: apellido
          type: string
        rol:
          $ref: "#/components/schemas/RolUsuario"
        condominioId:
          nullable: true
          title: condominioId
          type: integer
        activo:
          title: activo
          type: boolean
        fechaCreacion:
          format: date-time
          title: fechaCreacion
          type: string
        ultimoAcceso:
          format: date-time
          nullable: true
          title: ultimoAcceso
          type: string
      title: Usuario
      type: object
    UsuarioInput:
      example:
        password: password
        apellido: apellido
        condominioId: 0
        nombre: nombre
        email: email
        rol: SUPER_ADMINISTRADOR
      properties:
        email:
          format: email
          title: email
          type: string
        nombre:
          title: nombre
          type: string
        apellido:
          title: apellido
          type: string
        rol:
          $ref: "#/components/schemas/RolUsuario"
        condominioId:
          nullable: true
          title: condominioId
          type: integer
        password:
          format: password
          minLength: 8
          title: password
          type: string
      required:
      - apellido
      - email
      - nombre
      - password
      - rol
      title: UsuarioInput
      type: object
    RolUsuario:
      enum:
      - SUPER_ADMINISTRADOR
      - ADMINISTRADOR
      - CONSERJE
      - DIRECTIVA
      - RESIDENTE
      title: RolUsuario
      type: string
...
    _usuarios_get_200_response:
      example:
        data:
        - apellido: apellido
          fechaCreacion: 2000-01-23T04:56:07.000+00:00
          condominioId: 6
          id: 0
          ultimoAcceso: 2000-01-23T04:56:07.000+00:00
          nombre: nombre
          email: email
          rol: SUPER_ADMINISTRADOR
          activo: true
        - apellido: apellido
          fechaCreacion: 2000-01-23T04:56:07.000+00:00
          condominioId: 6
          id: 0
          ultimoAcceso: 2000-01-23T04:56:07.000+00:00
          nombre: nombre
          email: email
          rol: SUPER_ADMINISTRADOR
          activo: true
      properties:
        data:
          items:
            $ref: "#/components/schemas/Usuario"
          title: data
          type: array
      title: _usuarios_get_200_response
      type: object
...